requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
protobuf==3.12.2          # via -r requirements.in
pyhamcrest==2.0.2         # via -r requirements.in
six==1.15.0               # via protobuf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
